project1_AWS_Bedrock/requirements.txt
streamlit==1.28.1
boto3==1.34.0
Pillow==10.0.1
python-dotenv==1.0.0 